requirements.txt
streamlit==1.35.0
openai==1.34.0
langchain
langchain-openai
pypdf
python-dotenv